blockchain: change to a stricter reconstruct types
What's considered to be a reconstructable pkscript is now more strict. A
p2pkh sigscript may have extra opcodes that may break the assumption
that the sigscript will only have two data pushes of the signature and
the public key.

We redefine what's reconstructable as the last data push of the
sigscript must be the preimage of the pubkey hash or the script hash.
Anything that doesn't fit in this model is considered not
reconstructable.